requirements.txt
streamlit==1.36.0
python-dotenv==1.0.0
langchain==0.1.0
openai==1.3.5
pymongo==4.6.2
googletrans==3.1.0a0
plotly==5.18.0
gensim==4.3.2
textblob==0.17.1
python-docx==1.1.0
pdfkit==1.0.0
difflib36==0.2.3
google-search-results==2.4.2